+++ WriteOutputNetCDF.h
@@ -21,7 +21,7 @@ public:
   void initializeFile(const ProgramState*, const out_data_struct*);
   void openFile();
   void compressFiles();
-  void write_data_one_cell(out_data_struct *out_data, const dmy_struct *dmy, int numTimeRecords, const ProgramState* state);
+  void write_data_one_cell(out_data_struct *out_data, const dmy_struct *dmy, size_t numTimeRecords, const ProgramState* state);
   void write_data_all_cells(std::vector<out_data_struct*>& all_out_data, out_data_file_struct *out_data_files_template, const dmy_struct *dmy, int dt, const ProgramState *state);
   void write_header(out_data_struct *out_data, const dmy_struct *dmy, const ProgramState* state);
   int getLengthOfTimeDimension(const ProgramState* state);
